simplify by moving adapter to userdataschema
