--- collective/examples/userdata/adapter.py
@@ -1,15 +1,0 @@
-from .userdataschema import IEnhancedUserDataSchema
-from plone.app.users.browser.account import AccountPanelSchemaAdapter
-
-
-class EnhancedUserDataSchemaAdapter(AccountPanelSchemaAdapter):
-    schema = IEnhancedUserDataSchema
-
-    def get_birthdate(self):
-        bd = self._getProperty('birthdate')
-        return None if bd == '' else bd.asdatetime()
-
-    def set_birthdate(self, value):
-        return self._setProperty('birthdate', value)
-
-    birthdate = property(get_birthdate, set_birthdate)
