+++ collective/examples/userdata/userdataschema.py
@@ -7,12 +7,13 @@ from z3c.form import field
 from z3c.form.browser.radio import RadioFieldWidget
 
 from plone.supermodel import model
+from plone.app.users.browser.account import AccountPanelSchemaAdapter
 from plone.app.users.browser.userdatapanel import UserDataPanel
 from plone.app.users.browser.register import RegistrationForm, AddUserForm
 from plone.z3cform.fieldsets import extensible
 
-from collective.examples.userdata import _
 from collective.examples.userdata.interfaces import IUserDataExamplesLayer
+from collective.examples.userdata import _
 
 
 gender_options = SimpleVocabulary([
@@ -27,7 +28,7 @@ def validateAccept(value):
 
 
 class IEnhancedUserDataSchema(model.Schema):
-    """ Use all the fields from the default user data schema, and add various
+    """Use all the fields from the default user data schema, and add various
     extra fields.
     """
     firstname = schema.TextLine(
@@ -94,6 +95,19 @@ class IEnhancedUserDataSchema(model.Schema):
         required=True,
         constraint=validateAccept,
         )
+
+
+class EnhancedUserDataSchemaAdapter(AccountPanelSchemaAdapter):
+    schema = IEnhancedUserDataSchema
+
+    def get_birthdate(self):
+        bd = self._getProperty('birthdate')
+        return None if bd == '' else bd.asdatetime()
+
+    def set_birthdate(self, value):
+        return self._setProperty('birthdate', value)
+
+    birthdate = property(get_birthdate, set_birthdate)
 
 
 class UserDataPanelExtender(extensible.FormExtender):
